+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# (Unreleased)
+
+# Version 0.3.0
+
+## Added 
+
+- `playlist-list` command
+  - `-f` filter option
+  - `-s` sort option
 
 # 2022-12-18